Fix Barrack RouterId so the Zone Server starts at ID=0.

+++ server/R1EMU/src/Common/Redis/Fields/RedisGameSession.c
@@ -79,7 +79,7 @@ const char *redisGameSessionsStr [] = {
 bool
 Redis_requestSession (
     Redis *self,
-    int routerId,
+    uint16_t routerId,
     char *socketId,
     Session *session
 ) {
